# perf: Add performance logger to clean up