--- a/lib/BackgroundJob/CleanupJob.php
+++ b/lib/BackgroundJob/CleanupJob.php
@@ -28,20 +28,24 @@ namespace OCA\Mail\BackgroundJob;
 use OCA\Mail\Service\CleanupService;
 use OCP\AppFramework\Utility\ITimeFactory;
 use OCP\BackgroundJob\TimedJob;
+use Psr\Log\LoggerInterface;
 
 class CleanupJob extends TimedJob {
 	private CleanupService $cleanupService;
+	private LoggerInterface $logger;
 
 	public function __construct(ITimeFactory $time,
-		CleanupService $cleanupService) {
+		CleanupService $cleanupService,
+		LoggerInterface $logger) {
 		parent::__construct($time);
 		$this->cleanupService = $cleanupService;
+		$this->logger = $logger;
 
 		$this->setInterval(24 * 60 * 60);
 		$this->setTimeSensitivity(self::TIME_INSENSITIVE);
 	}
 
 	protected function run($argument): void {
-		$this->cleanupService->cleanUp();
+		$this->cleanupService->cleanUp($this->logger);
 	}
 }