--- a/lib/Service/CleanupService.php
+++ b/lib/Service/CleanupService.php
@@ -33,6 +33,9 @@ use OCA\Mail\Db\MessageRetentionMapper;
 use OCA\Mail\Db\MessageSnoozeMapper;
 use OCA\Mail\Db\TagMapper;
 use OCA\Mail\Service\Classification\PersistenceService;
+use OCA\Mail\Support\PerformanceLogger;
+use OCP\AppFramework\Utility\ITimeFactory;
+use Psr\Log\LoggerInterface;
 
 class CleanupService {
 	/** @var AliasMapper */
@@ -55,6 +58,7 @@ class CleanupService {
 	private MessageSnoozeMapper $messageSnoozeMapper;
 
 	private PersistenceService $classifierPersistenceService;
+	private ITimeFactory $timeFactory;
 
 	public function __construct(AliasMapper $aliasMapper,
 		MailboxMapper $mailboxMapper,
@@ -63,7 +67,8 @@ class CleanupService {
 		TagMapper $tagMapper,
 		MessageRetentionMapper $messageRetentionMapper,
 		MessageSnoozeMapper $messageSnoozeMapper,
-		PersistenceService $classifierPersistenceService) {
+		PersistenceService $classifierPersistenceService,
+		ITimeFactory $timeFactory) {
 		$this->aliasMapper = $aliasMapper;
 		$this->mailboxMapper = $mailboxMapper;
 		$this->messageMapper = $messageMapper;
@@ -72,17 +77,32 @@ class CleanupService {
 		$this->messageRetentionMapper = $messageRetentionMapper;
 		$this->messageSnoozeMapper = $messageSnoozeMapper;
 		$this->classifierPersistenceService = $classifierPersistenceService;
+		$this->timeFactory = $timeFactory;
 	}
 
-	public function cleanUp(): void {
+	public function cleanUp(LoggerInterface $logger): void {
+		$task = (new PerformanceLogger(
+			$this->timeFactory,
+			$logger
+		))->start('clean up');
 		$this->aliasMapper->deleteOrphans();
+		$task->step('delete orphan aliases');
 		$this->mailboxMapper->deleteOrphans();
+		$task->step('delete orphan mailboxes');
 		$this->messageMapper->deleteOrphans();
+		$task->step('delete orphan messages');
 		$this->collectedAddressMapper->deleteOrphans();
+		$task->step('delete orphan collected addresses');
 		$this->tagMapper->deleteOrphans();
+		$task->step('delete orphan tags');
 		$this->tagMapper->deleteDuplicates();
+		$task->step('delete duplicate tags');
 		$this->messageRetentionMapper->deleteOrphans();
+		$task->step('delete expired messages');
 		$this->messageSnoozeMapper->deleteOrphans();
+		$task->step('delete orphan snoozes');
 		$this->classifierPersistenceService->cleanUp();
+		$task->step('delete orphan classifiers');
+		$task->end();
 	}
 }