Helps gather data for https://github.com/nextcloud/mail/issues/9079.

```
$ php occ mail:clean-up -vvv
[debug] clean up - delete orphan aliases took 0s. 10/10MB memory used
[debug] clean up - delete orphan mailboxes took 0s. 10/10MB memory used
[debug] clean up - delete orphan messages took 0s. 10/10MB memory used
[debug] clean up - delete orphan collected addresses took 0s. 10/10MB memory used
[debug] clean up - delete orphan tags took 0s. 10/10MB memory used
[debug] clean up - delete duplicate tags took 0s. 10/10MB memory used
[debug] clean up - delete expired messages took 0s. 10/10MB memory used
[debug] clean up - delete orphan snoozes took 0s. 10/10MB memory used
[debug] clean up - delete orphan classifiers took 1s. 11/11MB memory used
[debug] clean up took 1s
```