@ChristophWurst 2023-11-22
Code coverage drops here but I'd prefer skipping tests for this.