@ChristophWurst 2023-11-29
/backport to stable3.4